First commit in Bootcamp-T-Systems
In this project, we were challenged to develop a medical consultation system in Java. The system allows users to:

Register a client

Calculate and display the Body Mass Index (BMI)

Start a consultation

Schedule a consultation

Display client data

Exit the system

We used a while-do loop to manage the main menu flow, and implemented object-oriented principles through classes such as App and Consulta. We also used a switch-case structure for menu navigation. The project was developed using JetBrains IntelliJ IDEA.